refactor: extract filesystem helpers from AgentWorkspace
